Issue #32 Fix issues with GetContactScoresCommand and GetScoringModelsCommand     - GetContactScoresCommand#contactId should be a Long, not Integer     - com.silverpop.api.client.converters.ApiDateConverter works only with one fixed format, fails for others     - typo in the annotation of GetContactScoresModel#scoreComponents     - GetContactScoresModelScoreComponent#value could be a String, not only an Integer     - the annotation value for GetScoringModelsModel#scoreComponents should be SCORE_COMPONENT_TYPES, not SCORE_COMPONENTS     - the annotation value for GetScoringModelsModelScoreComponent should be SCORE_COMPONENT_TYPE, not SCORE_COMPONENT     - missing annotation of GetScoringModelsModelScoreComponentValues

+++ src/main/java/com/silverpop/api/client/command/GetContactScoresCommand.java
@@ -11,18 +11,18 @@ import com.silverpop.api.client.result.GetContactScoresResult;
 public class GetContactScoresCommand implements ApiCommand {
 	
 	@XStreamAlias("CONTACT_ID")
-	private Integer contactId;
+	private Long contactId;
 
 	@Override
 	public Class<? extends ApiResult> getResultType() {
 		return GetContactScoresResult.class;
 	}
 	
-	public void setContactId(Integer contactId) {
+	public void setContactId(Long contactId) {
 		this.contactId = contactId;
 	}
 	
-	public Integer getContactId() {
+	public Long getContactId() {
 		return contactId;
 	}
 }
